APS项目后端项目完成所有基础表的增删改查

+++ demo-api/src/main/java/com/xk/aps/service/impl/XkApsAttendanceServiceImpl.java
@@ -4,16 +4,23 @@ import com.xk.aps.dao.XkApsAttendanceRepository;
 import com.xk.aps.model.dto.XkApsAttendanceDto;
 import com.xk.aps.model.entity.XkApsAttendanceEntity;
 import com.xk.framework.common.*;
+import com.xk.framework.jpa.specification.SimpleSpecificationBuilder;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import com.xk.aps.service.IXkApsAttendanceService;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 /**
 * 描述：一键生成单表模块服务实现层
@@ -33,6 +40,7 @@ public class XkApsAttendanceServiceImpl implements IXkApsAttendanceService {
     * @param formData 保存或者更新的实体
     * @return XkApsAttendanceDto
     */
+    @CacheEvict(value = "Attendance",allEntries = true)
     @Override
     @Transactional
     public XkApsAttendanceDto save(XkApsAttendanceDto formData) {
@@ -58,6 +66,7 @@ public class XkApsAttendanceServiceImpl implements IXkApsAttendanceService {
     * @param id 主键id
     * @return XkApsAttendanceDto
     */
+    @CacheEvict(value = "Attendance",allEntries = true)
     @Override
     @Transactional(readOnly = true)
     public XkApsAttendanceDto getById(String id) {
@@ -78,11 +87,15 @@ public class XkApsAttendanceServiceImpl implements IXkApsAttendanceService {
     *
     * @param ids 删除的id
     */
+    @CacheEvict(value = "Attendance",allEntries = true)
     @Override
-    @Transactional
+    @Transactional(rollbackFor = Exception.class)
     public void removeMulti(String ids) {
+        System.out.println("方法中"+ids);
         try {
-            xkApsAttendanceRepository.removeMuiltByIds(ids);
+            if(!ids.equals(null) && !ids.equals("")){
+                xkApsAttendanceRepository.removeMuiltByIds(ids);
+            }
         } catch (Exception e) {
             logger.error("删除数据失败,{}",e);
             e.printStackTrace();
@@ -95,8 +108,9 @@ public class XkApsAttendanceServiceImpl implements IXkApsAttendanceService {
     *
     * @param id 一键生成单表模块编号
     */
+    @CacheEvict(value = "Attendance",allEntries = true)
     @Override
-    @Transactional
+    @Transactional(rollbackFor = Exception.class)
     public void remove(String id) {
         try {
              xkApsAttendanceRepository.remove(id);
@@ -106,6 +120,70 @@ public class XkApsAttendanceServiceImpl implements IXkApsAttendanceService {
             throw new ServiceException("删除数据失败", CommonErrorCode.DELETE_ERROR);
         }
     }
+
+    /**
+     * 获取出勤表所有信息
+     * @return PageDto {@code<XkApsAttendanceDto>}
+     */
+    //@Cacheable(value = {"Attendance"}, key = "#root.method.name",sync = true)
+    @Override
+    public List<XkApsAttendanceDto> listAll(@RequestParam Map<String, Object> params) {
+        try {
+            String key = (String) params.get("key");
+            List<XkApsAttendanceEntity> attendanceEntities;
+
+            //自定义条件查询
+            SimpleSpecificationBuilder<XkApsAttendanceEntity> ssb=new SimpleSpecificationBuilder<XkApsAttendanceEntity>();
+            if(!StringUtils.isEmpty(key) || key != ""){
+                ssb.add("delFlag", Constants.OPER_EQ, 0);
+                ssb.add("attendanceCode", Constants.OPER_LIKE, key);
+                ssb.addOr("id", Constants.OPER_LIKE, key);
+                attendanceEntities = xkApsAttendanceRepository.findAll(ssb.generateSpecification());
+            }else {
+                ssb.add("delFlag", Constants.OPER_EQ, 0);
+                attendanceEntities = xkApsAttendanceRepository.findAll(ssb.generateSpecification());
+            }
+            List<XkApsAttendanceDto> collect = attendanceEntities.stream().map((item) -> {
+                XkApsAttendanceDto xkApsAttendanceDto = BeanMapperUtils.map(item, XkApsAttendanceDto.class);
+                return xkApsAttendanceDto;
+            }).collect(Collectors.toList());
+
+
+            return collect;
+
+
+        } catch (Exception e) {
+            logger.error("查询所有数据失败,{}",e);
+            e.printStackTrace();
+            throw new ServiceException("查询所有数据失败", CommonErrorCode.SELECT_ERROR);
+        }
+    }
+    /**
+     * 删除多个出勤表模块信息
+     * @param ids 删除的id
+     */
+//    @Cacheable(value = {"Attendance"}, key = "#root.method.name",sync = true)
+//    @Transactional(rollbackFor = Exception.class)
+//    @Override
+//    public void deleteIds(List<String> ids) {
+//        try {
+//            System.out.println("开始删除方法");
+//            if(ids!=null && ids.size()>0){
+////                ids.stream().map((item)->{
+////                    xkApsAttendanceRepository.remove(item);
+////                    return null;
+////                });
+//                for(String id : ids){
+//                    xkApsAttendanceRepository.deleteIds(id);
+//                }
+//                //xkApsAttendanceRepository.remove();
+//            }
+//        } catch (Exception e) {
+//            logger.error("删除数据失败,{}",e);
+//            e.printStackTrace();
+//            throw new ServiceException("删除数据失败", CommonErrorCode.DELETE_ERROR);
+//        }
+//    }
 
     /**
     * 分页获取一键生成单表模块信息
